Moved time duration formatter to component folder.
