+++ console/src/components/MapCore/MapCore.js
@@ -15,46 +15,12 @@ import AppContextConnect from '../../App/AppContextConnect';
 import appConfig from '../../App/configLoader';
 import {propTypeLatLon, propTypeLatLonList} from '../../data/proptypes';
 import CarPng from '../Dashboard/svg/car.png';
+import {formatDuration, formatTime} from '../../../../components/Formatter';
 
 import css from './MapCore.less';
 
 const linear = (input) => input;
-const zeroPad = (val) => (val < 10 ? '0' + val : val);
 
-const formatTime = (time) => {
-	const formattedEta = new Date(time);
-	const hour = formattedEta.getHours() % 12 || 12,
-		min = zeroPad(formattedEta.getMinutes()),
-		// sec = zeroPad(formattedEta.getSeconds()),
-		ampm = (formattedEta.getHours() >= 13 ? 'PM' : 'AM');
-	return `${hour}:${min} ${ampm}`;
-};
-
-// This array maps 1:1 to the durValues array below
-const durationIncrements = ['day', 'hour', 'min'];
-const formatDuration = (duration) => {
-	duration = Math.ceil(duration);
-	const durValues = [
-		Math.floor(duration / (60 * 60 * 24)),  // lol we can stop at days
-		Math.floor(duration / (60 * 60)) % 24,
-		Math.floor(duration / 60) % 60
-	];
-
-	// It's only useful to show the two largest increments
-	const durParts = [];
-	for (let i = 0, useful = 0; i < durValues.length && useful < 2; i++) {
-		if (durValues[i] || useful) {
-			useful++;
-		}
-		// `zero` values are not displayed, but still counted as useful
-		if (durValues[i]) {
-			// stack up the number, unit, and pluralize the unit
-			durParts[i] = durValues[i] + ' ' + durationIncrements[i] + (durValues[i] === 1 ? '' : 's');
-		}
-	}
-	// Prune the empty ones and join the rest.
-	return durParts.filter(part => !!part).join(' ');
-};
 
 if (!appConfig.mapApiKey) {
 	Error('Please set `mapApiKey` key in your `config.js` file to your own Mapbox API key.');
@@ -585,6 +551,7 @@ class MapCoreBase extends React.Component {
 		delete rest.viewLockoutDuration;
 		delete rest.zoomToSpeedScaleFactor;
 		const {duration, distance, eta, selectedDestination, selfDriving} = this.state;
+		const durationIncrements = ['day', 'hour', 'min'];
 
 		return (
 			<div {...rest} className={classnames(className, css.map)}>
@@ -633,7 +600,7 @@ class MapCoreBase extends React.Component {
 					{
 						selectedDestination &&
 						<div>
-							<p>{formatDuration(duration)}</p>
+							<p>{formatDuration(duration, durationIncrements)}</p>
 							<p>{(distance / 1609.344).toFixed(1)} mi - {formatTime(eta)}</p>
 							<ToggleButton
 								className={css.button}
